+++ packages/@eventual/core/src/runtime/clients/execution-history-client.ts
@@ -1,4 +1,4 @@
-import { WorkflowEvent } from "../events.js";
+import { WorkflowEvent } from "../../events.js";
 
 export type UnresolvedEvent<T extends WorkflowEvent> = Omit<
   T,
