+++ packages/@eventual/core/src/runtime/clients/workflow-runtime-client.ts
@@ -3,14 +3,14 @@ import {
   SleepForCommand,
   SleepUntilCommand,
   ExpectSignalCommand,
-} from "../command.js";
+} from "../../command.js";
 import {
   ActivityScheduled,
   HistoryStateEvent,
   SleepScheduled,
   ExpectSignalStarted,
-} from "../events.js";
-import { CompleteExecution, FailedExecution } from "../execution.js";
+} from "../../events.js";
+import { CompleteExecution, FailedExecution } from "../../execution.js";
 
 export interface CompleteExecutionRequest {
   executionId: string;
