+++ packages/@eventual/core/src/runtime/index.ts
@@ -1,6 +1,2 @@
-export * from "./activity-runtime-client.js";
-export * from "./execution-history-client.js";
-export * from "./timer-client.js";
-export * from "./workflow-client.js";
-export * from "./workflow-runtime-client.js";
+export * from "./clients/index.js";
 export * from "./flags.js";
